refactor: decouple public routes from heavy providers
