+++ src/frontend/src/clerk/auth.tsx
@@ -1,4 +1,4 @@
-import { lazy, ReactNode, useContext, useEffect, useRef } from "react";
+import { ReactNode, useContext, useEffect, useRef } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import { AuthContext } from "@/contexts/authContext";
 import { api } from "@/controllers/API/api";
@@ -8,14 +8,12 @@ import { ClerkProvider, useAuth, useClerk, useOrganization, useUser, SignedOut }
 import { Users } from "@/types/api";
 import { LANGFLOW_ACCESS_TOKEN, LANGFLOW_REFRESH_TOKEN } from "@/constants/constants";
 import { Cookies } from "react-cookie";
-import OrganizationPage from "./OrganizationPage";
 import authStore from "@/stores/authStore";
-
-export const IS_CLERK_AUTH =
-  String(import.meta.env.VITE_CLERK_AUTH_ENABLED).toLowerCase() === "true";
-
-export const CLERK_PUBLISHABLE_KEY = import.meta.env.VITE_CLERK_PUBLISHABLE_KEY || "";
-export const CLERK_DUMMY_PASSWORD = "clerk_dummy_password";
+import {
+  CLERK_DUMMY_PASSWORD,
+  CLERK_PUBLISHABLE_KEY,
+  IS_CLERK_AUTH,
+} from "@/clerk/config";
 
 export enum HttpStatusCode {
   UNAUTHORIZED = 401,
