+++ src/frontend/src/controllers/API/queries/auth/use-get-autologin.ts
@@ -13,7 +13,7 @@ import { api } from "../../api";
 import { getURL } from "../../helpers/constants";
 import { UseRequestProcessor } from "../../services/request-processor";
 import { useLogout } from "./use-post-logout";
-import { IS_CLERK_AUTH } from "@/clerk/auth";
+import { IS_CLERK_AUTH } from "@/clerk/config";
 
 export interface AutoLoginResponse {
   frontend_timeout: number;
