+++ src/frontend/src/routes.tsx
@@ -7,9 +7,7 @@ import {
 } from "react-router-dom";
 import { ProtectedAdminRoute } from "./components/authorization/authAdminGuard";
 import { ProtectedRoute } from "./components/authorization/authGuard";
-import { ProtectedLoginRoute } from "./components/authorization/authLoginGuard";
 import { AuthSettingsGuard } from "./components/authorization/authSettingsGuard";
-import ContextWrapper from "./contexts";
 import { CustomNavigate } from "./customization/components/custom-navigate";
 import { BASENAME } from "./customization/config-constants";
 import {
@@ -18,14 +16,17 @@ import {
 } from "./customization/feature-flags";
 import { CustomRoutesStore } from "./customization/utils/custom-routes-store";
 import { CustomRoutesStorePages } from "./customization/utils/custom-routes-store-pages";
-import { IS_CLERK_AUTH } from "./clerk/auth";
+import { IS_CLERK_AUTH } from "./clerk/config";
 import { LoadingPage } from "./pages/LoadingPage";
+import useAuthStore from "./stores/authStore";
 
 const AppWrapperPage = lazy(() =>
   import("./pages/AppWrapperPage").then((module) => ({
     default: module.AppWrapperPage,
   })),
 );
+const LandingPage = lazy(() => import("./pages/LandingPage"));
+const ContextWrapper = lazy(() => import("./contexts"));
 const AppInitPage = lazy(() =>
   import("./pages/AppInitPage").then((module) => ({
     default: module.AppInitPage,
@@ -75,9 +76,30 @@ const LoginAdminPage = lazy(() =>
   })),
 );
 
+const ProtectedLoginRoute = lazy(() =>
+  import("./components/authorization/authLoginGuard").then((module) => ({
+    default: module.ProtectedLoginRoute,
+  })),
+);
+
 const AdminPage = lazy(() => import("./pages/AdminPage"));
 const DeleteAccountPage = lazy(() => import("./pages/DeleteAccountPage"));
 const PlaygroundPage = lazy(() => import("./pages/Playground"));
+
+function LandingEntryRoute() {
+  const isAuthenticated = useAuthStore((state) => state.isAuthenticated);
+
+  if (isAuthenticated) {
+    const target = IS_CLERK_AUTH ? "/organization" : "flows";
+    return <CustomNavigate replace to={target} />;
+  }
+
+  return (
+    <Suspense fallback={<LoadingPage />}>
+      <LandingPage />
+    </Suspense>
+  );
+}
 
 const router = createBrowserRouter(
   createRoutesFromElements([
@@ -85,27 +107,65 @@ const router = createBrowserRouter(
       <Route
         path=""
         element={
-          <ContextWrapper key={1}>
-            <Suspense fallback={<LoadingPage />}>
+          <Suspense fallback={<LoadingPage />}>
+            <ContextWrapper key={1}>
               <PlaygroundPage />
-            </Suspense>
-          </ContextWrapper>
+            </ContextWrapper>
+          </Suspense>
         }
       />
     </Route>,
     <Route
       path={ENABLE_CUSTOM_PARAM ? "/:customParam?" : "/"}
-      element={
-        <ContextWrapper key={2}>
-          <Outlet />
-        </ContextWrapper>
-      }
+      element={<Outlet />}
     >
+      <Route index element={<LandingEntryRoute />} />
       <Route
-        path=""
+        path="login"
         element={
           <Suspense fallback={<LoadingPage />}>
-            <AppWrapperPage />
+            <ProtectedLoginRoute>
+              <LoginPage />
+            </ProtectedLoginRoute>
+          </Suspense>
+        }
+      />
+      <Route
+        path="organization"
+        element={
+          <Suspense fallback={<LoadingPage />}>
+            <ProtectedLoginRoute>
+              <OrganizationPage />
+            </ProtectedLoginRoute>
+          </Suspense>
+        }
+      />
+      <Route
+        path="signup"
+        element={
+          <Suspense fallback={<LoadingPage />}>
+            <ProtectedLoginRoute>
+              <SignUp />
+            </ProtectedLoginRoute>
+          </Suspense>
+        }
+      />
+      <Route
+        path="login/admin"
+        element={
+          <Suspense fallback={<LoadingPage />}>
+            <ProtectedLoginRoute>
+              <LoginAdminPage />
+            </ProtectedLoginRoute>
+          </Suspense>
+        }
+      />
+      <Route
+        element={
+          <Suspense fallback={<LoadingPage />}>
+            <ContextWrapper key={2}>
+              <Outlet />
+            </ContextWrapper>
           </Suspense>
         }
       >
@@ -113,9 +173,7 @@ const router = createBrowserRouter(
           path=""
           element={
             <Suspense fallback={<LoadingPage />}>
-              <ProtectedRoute>
-                <AppInitPage />
-              </ProtectedRoute>
+              <AppWrapperPage />
             </Suspense>
           }
         >
@@ -123,7 +181,9 @@ const router = createBrowserRouter(
             path=""
             element={
               <Suspense fallback={<LoadingPage />}>
-                <AppAuthenticatedPage />
+                <ProtectedRoute>
+                  <AppInitPage />
+                </ProtectedRoute>
               </Suspense>
             }
           >
@@ -131,174 +191,10 @@ const router = createBrowserRouter(
               path=""
               element={
                 <Suspense fallback={<LoadingPage />}>
-                  <CustomDashboardWrapperPage />
+                  <AppAuthenticatedPage />
                 </Suspense>
               }
             >
-              <Route
-                path=""
-                element={
-                  <Suspense fallback={<LoadingPage />}>
-                    <CollectionPage />
-                  </Suspense>
-                }
-              >
-                <Route
-                  index
-                  element={
-                    IS_CLERK_AUTH ? (
-                      <CustomNavigate replace to="/organization" />
-                    ) : (
-                      <CustomNavigate replace to="flows" />
-                    )
-                  }
-                />
-                {ENABLE_FILE_MANAGEMENT && (
-                  <Route
-                    path="files"
-                    element={
-                      <Suspense fallback={<LoadingPage />}>
-                        <FilesPage />
-                      </Suspense>
-                    }
-                  />
-                )}
-                <Route
-                  path="flows/"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <HomePage key="flows" type="flows" />
-                    </Suspense>
-                  }
-                />
-                <Route
-                  path="components/"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <HomePage key="components" type="components" />
-                    </Suspense>
-                  }
-                >
-                  <Route
-                    path="folder/:folderId"
-                    element={
-                      <Suspense fallback={<LoadingPage />}>
-                        <HomePage key="components" type="components" />
-                      </Suspense>
-                    }
-                  />
-                </Route>
-                <Route
-                  path="all/"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <HomePage key="flows" type="flows" />
-                    </Suspense>
-                  }
-                >
-                  <Route
-                    path="folder/:folderId"
-                    element={
-                      <Suspense fallback={<LoadingPage />}>
-                        <HomePage key="flows" type="flows" />
-                      </Suspense>
-                    }
-                  />
-                </Route>
-                <Route
-                  path="mcp/"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <HomePage key="mcp" type="mcp" />
-                    </Suspense>
-                  }
-                >
-                  <Route
-                    path="folder/:folderId"
-                    element={
-                      <Suspense fallback={<LoadingPage />}>
-                        <HomePage key="mcp" type="mcp" />
-                      </Suspense>
-                    }
-                  />
-                </Route>
-              </Route>
-              <Route
-                path="settings"
-                element={
-                  <Suspense fallback={<LoadingPage />}>
-                    <SettingsPage />
-                  </Suspense>
-                }
-              >
-                <Route index element={<CustomNavigate replace to="general" />} />
-                <Route
-                  path="global-variables"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <GlobalVariablesPage />
-                    </Suspense>
-                  }
-                />
-                <Route
-                  path="api-keys"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <ApiKeysPage />
-                    </Suspense>
-                  }
-                />
-                <Route
-                  path="general/:scrollId?"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <AuthSettingsGuard>
-                        <GeneralPage />
-                      </AuthSettingsGuard>
-                    </Suspense>
-                  }
-                />
-                <Route
-                  path="shortcuts"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <ShortcutsPage />
-                    </Suspense>
-                  }
-                />
-                <Route
-                  path="messages"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <MessagesPage />
-                    </Suspense>
-                  }
-                />
-                {CustomRoutesStore()}
-              </Route>
-              {CustomRoutesStorePages()}
-              <Route path="account">
-                <Route
-                  path="delete"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <DeleteAccountPage />
-                    </Suspense>
-                  }
-                />
-              </Route>
-              <Route
-                path="admin"
-                element={
-                  <Suspense fallback={<LoadingPage />}>
-                    <ProtectedAdminRoute>
-                      <AdminPage />
-                    </ProtectedAdminRoute>
-                  </Suspense>
-                }
-              />
-            </Route>
-            <Route path="flow/:id/">
               <Route
                 path=""
                 element={
@@ -308,76 +204,209 @@ const router = createBrowserRouter(
                 }
               >
                 <Route
-                  path="folder/:folderId/"
-                  element={
-                    <Suspense fallback={<LoadingPage />}>
-                      <FlowPage />
-                    </Suspense>
-                  }
-                />
-                <Route
                   path=""
                   element={
                     <Suspense fallback={<LoadingPage />}>
-                      <FlowPage />
+                      <CollectionPage />
+                    </Suspense>
+                  }
+                >
+                  <Route
+                    index
+                    element={
+                      IS_CLERK_AUTH ? (
+                        <CustomNavigate replace to="/organization" />
+                      ) : (
+                        <CustomNavigate replace to="flows" />
+                      )
+                    }
+                  />
+                  {ENABLE_FILE_MANAGEMENT && (
+                    <Route
+                      path="files"
+                      element={
+                        <Suspense fallback={<LoadingPage />}>
+                          <FilesPage />
+                        </Suspense>
+                      }
+                    />
+                  )}
+                  <Route
+                    path="flows/"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <HomePage key="flows" type="flows" />
+                      </Suspense>
+                    }
+                  />
+                  <Route
+                    path="components/"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <HomePage key="components" type="components" />
+                      </Suspense>
+                    }
+                  >
+                    <Route
+                      path="folder/:folderId"
+                      element={
+                        <Suspense fallback={<LoadingPage />}>
+                          <HomePage key="components" type="components" />
+                        </Suspense>
+                      }
+                    />
+                  </Route>
+                  <Route
+                    path="all/"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <HomePage key="flows" type="flows" />
+                      </Suspense>
+                    }
+                  >
+                    <Route
+                      path="folder/:folderId"
+                      element={
+                        <Suspense fallback={<LoadingPage />}>
+                          <HomePage key="flows" type="flows" />
+                        </Suspense>
+                      }
+                    />
+                  </Route>
+                  <Route
+                    path="mcp/"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <HomePage key="mcp" type="mcp" />
+                      </Suspense>
+                    }
+                  >
+                    <Route
+                      path="folder/:folderId"
+                      element={
+                        <Suspense fallback={<LoadingPage />}>
+                          <HomePage key="mcp" type="mcp" />
+                        </Suspense>
+                      }
+                    />
+                  </Route>
+                </Route>
+                <Route
+                  path="settings"
+                  element={
+                    <Suspense fallback={<LoadingPage />}>
+                      <SettingsPage />
+                    </Suspense>
+                  }
+                >
+                  <Route index element={<CustomNavigate replace to="general" />} />
+                  <Route
+                    path="global-variables"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <GlobalVariablesPage />
+                      </Suspense>
+                    }
+                  />
+                  <Route
+                    path="api-keys"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <ApiKeysPage />
+                      </Suspense>
+                    }
+                  />
+                  <Route
+                    path="general/:scrollId?"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <AuthSettingsGuard>
+                          <GeneralPage />
+                        </AuthSettingsGuard>
+                      </Suspense>
+                    }
+                  />
+                  <Route
+                    path="shortcuts"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <ShortcutsPage />
+                      </Suspense>
+                    }
+                  />
+                  <Route
+                    path="messages"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <MessagesPage />
+                      </Suspense>
+                    }
+                  />
+                  {CustomRoutesStore()}
+                </Route>
+                {CustomRoutesStorePages()}
+                <Route path="account">
+                  <Route
+                    path="delete"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <DeleteAccountPage />
+                      </Suspense>
+                    }
+                  />
+                </Route>
+                <Route
+                  path="admin"
+                  element={
+                    <Suspense fallback={<LoadingPage />}>
+                      <ProtectedAdminRoute>
+                        <AdminPage />
+                      </ProtectedAdminRoute>
                     </Suspense>
                   }
                 />
               </Route>
-              <Route
-                path="view"
-                element={
-                  <Suspense fallback={<LoadingPage />}>
-                    <ViewPage />
-                  </Suspense>
-                }
-              />
+              <Route path="flow/:id/">
+                <Route
+                  path=""
+                  element={
+                    <Suspense fallback={<LoadingPage />}>
+                      <CustomDashboardWrapperPage />
+                    </Suspense>
+                  }
+                >
+                  <Route
+                    path="folder/:folderId/"
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <FlowPage />
+                      </Suspense>
+                    }
+                  />
+                  <Route
+                    path=""
+                    element={
+                      <Suspense fallback={<LoadingPage />}>
+                        <FlowPage />
+                      </Suspense>
+                    }
+                  />
+                </Route>
+                <Route
+                  path="view"
+                  element={
+                    <Suspense fallback={<LoadingPage />}>
+                      <ViewPage />
+                    </Suspense>
+                  }
+                />
+              </Route>
             </Route>
           </Route>
         </Route>
-        <Route
-          path="login"
-          element={
-            <Suspense fallback={<LoadingPage />}>
-              <ProtectedLoginRoute>
-                <LoginPage />
-              </ProtectedLoginRoute>
-            </Suspense>
-          }
-        />
-        <Route
-          path="organization"
-          element={
-            <Suspense fallback={<LoadingPage />}>
-              <ProtectedLoginRoute>
-                <OrganizationPage />
-              </ProtectedLoginRoute>
-            </Suspense>
-          }
-        />
-        <Route
-          path="signup"
-          element={
-            <Suspense fallback={<LoadingPage />}>
-              <ProtectedLoginRoute>
-                <SignUp />
-              </ProtectedLoginRoute>
-            </Suspense>
-          }
-        />
-        <Route
-          path="login/admin"
-          element={
-            <Suspense fallback={<LoadingPage />}>
-              <ProtectedLoginRoute>
-                <LoginAdminPage />
-              </ProtectedLoginRoute>
-            </Suspense>
-          }
-        />
       </Route>
-      <Route path="*" element={<CustomNavigate replace to="/" />} />
     </Route>,
+    <Route path="*" element={<CustomNavigate replace to="/" />} />,
   ]),
   { basename: BASENAME || undefined },
 );
